refactor(minishell): use size_t for indices and counts for consistency
Change int to size_t for variables representing sizes or counts to
ensure consistency and prevent potential overflow issues.

refactor(minishell): use const for string parameters to enhance safety

Add const qualifier to string parameters where applicable to indicate
that the function does not modify the input string, improving code
safety and readability.

fix(minishell): correct loop initialization and increment for clarity

Adjust loop initialization and increment to improve code clarity and
prevent potential off-by-one errors.

fix(minishell): improve error handling in memory allocation

Refactor error handling in memory allocation to ensure proper cleanup
and error reporting, preventing potential memory leaks.

+++ src/environment/env.c
@@ -6,7 +6,7 @@
 /*   By: ladloff <ladloff@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/27 12:42:17 by ladloff           #+#    #+#             */
-/*   Updated: 2024/04/16 15:21:14 by ladloff          ###   ########.fr       */
+/*   Updated: 2024/11/05 18:46:33 by ladloff          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -119,7 +119,7 @@ int	manage_environment(t_env_list **env)
 	while (*environ)
 	{
 		equals_location = ft_strchr(*environ, '=');
-		name = ft_strndup(*environ, equals_location - *environ);
+		name = ft_strndup(*environ, (size_t)(equals_location - *environ));
 		if (!name)
 			return (perror("ft_strndup (manage_environment)"), 1);
 		value = ft_strdup(equals_location + 1);
